chore(tasks): update tasks header
